soc-tsc-mapper/requirements.txt
fastapi
uvicorn
numpy
faiss-cpu
pydantic
openai>=1.0.0
